backend/requirements.txt
asgiref==3.8.1
contourpy==1.3.1
cycler==0.12.1
DateTime==5.5
Django==5.1.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Faker==37.1.0
fonttools==4.56.0
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.3
packaging==24.2
pillow==11.1.0
PyJWT==2.9.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
zope.interface==7.2